+++ BattleUnits/AlienSoldier.h
@@ -5,7 +5,7 @@
 
 class AlienSoldier : public Unit {
    public:
-    AlienSoldier(Game* game, int joinTime, int health, int power, int attackCapacity);
+    AlienSoldier(Game* game, int joinTime, double health, double power, int attackCapacity);
     void attack() override;
 };
